Added the completed program
Each time the user makes an input the buzzer will emit a sound to indicate to the user that an input was registered. If the password attempt is correct the buzzer will emit a low buzz followed by a high buzz which is then followed by the red LED lighting up. After this, the lock will wait for the user to press any button again before locking it and turning off the LED. After locking the device, it will wait for more inputs from the user. If the user inputs an incorrect combination a high buzz will emit and be followed up by a low buzz. This indicates that the combination was incorrect. The user may proceed to continue their attempts after the buzz. Implemented Pushbuttons, Piezo Buzzers, a single LED, a single 470-ohm resistor, an MPLAB SNAP, and an ATtiny2313A.

Edits: There are multiple loops within the code that should've been turned into functions. This would've greatly improved readability.